Use multiple inheritance for BondSpline and AngleSpline.
